Fix remaining compile errors in torture.pony.c
Since we don't really have a great way of saying "this is a statement", we have to be a little messy and do our best to try and make sure a ; gets inserted at the right place.

We need to do the same thing for while and for and other things that could take a single statement or a group. OTOH we could enforce, in the parser, that that node is always a Group. This would make pattern matching easier but make construction more obtuse. Sigh.